+++ Public/Timeline/New-MermaidTimelineDirective.ps1
@@ -1,0 +1,21 @@
+function New-MermaidTimelineDirective {
+    [CmdletBinding()]
+    param (
+        [ValidateSet("base","forest","dark","default","neutral")]
+        $Theme,
+        [switch] $DisableMulticolor
+    )
+    
+    begin {
+        $ReturnString = ""
+    }
+    
+    process {
+        $ReturnString = (("%%{init: {'logLevel': 'debug', 'theme': 'RPTHEME', 'timeline': {'disableMulticolor': RPDISABLEMULTICOLOR}}}%%").Replace('RPTHEME',$Theme)).Replace('RPDISABLEMULTICOLOR',$DisableMulticolor)
+        
+    }
+    
+    end {
+        return $ReturnString
+    }
+}